requirements.txt
numpy==1.24.4
matplotlib~=3.7.5
SimpleITK~=2.4.1
MedPy~=0.5.2
scipy~=1.10.1
pillow~=10.4.0
ptflops==0.7.2.2
torchvision~=0.12.0+cu113
h5py~=3.11.0
opencv-python~=4.11.0.86
tqdm~=4.67.1
tensorboardX~=2.6.2.2
timm~=1.0.12